v17
added group implementation, control header.
